+++ acts/src/store/tests/mem.rs
@@ -132,7 +132,7 @@ async fn store_mem_proc_create() {
         end_time: 0,
         timestamp: 0,
         model: "".to_string(),
-        env_local: "".to_string(),
+        env: "".to_string(),
         err: None,
     };
     store.procs().create(&proc).unwrap();
@@ -152,7 +152,7 @@ async fn store_mem_proc_find() {
         end_time: 0,
         timestamp: 0,
         model: "".to_string(),
-        env_local: "".to_string(),
+        env: "".to_string(),
         err: None,
     };
     store.procs().create(&proc).unwrap();
@@ -174,7 +174,7 @@ async fn store_mem_proc_query() {
             end_time: 0,
             timestamp: 0,
             model: "".to_string(),
-            env_local: "".to_string(),
+            env: "".to_string(),
             err: None,
         };
         procs.create(&proc).unwrap();
@@ -203,7 +203,7 @@ async fn store_mem_proc_update() {
         end_time: 0,
         timestamp: 0,
         model: "".to_string(),
-        env_local: "".to_string(),
+        env: "".to_string(),
         err: None,
     };
     store.procs().create(&proc).unwrap();
@@ -227,7 +227,7 @@ async fn store_mem_proc_delete() {
         end_time: 0,
         timestamp: 0,
         model: "".to_string(),
-        env_local: "".to_string(),
+        env: "".to_string(),
         err: None,
     };
     store.procs().create(&proc).unwrap();